initial
This is a check-in just more or less the Visual Studio WinUI 3 app template with some corrections and edits to reduce binary size and remove unnecessary dependencies. I can see that the visual designer value add for these apps is pretty low and the thing is going to be replaced in the next iteration, it's actually nicer editing xaml as text, so I threw away VS and am working in VSCode.